requirements.txt
strealit
pandas
plotly_express
openpyxl